Create videoprocessor.py
This is the main image processing script. It fetches frames from the video feed using gstreamer and passes it to OpenCV2. We then extract the white-ish pixels corresponding to the pipe, use canny and probabilistic hough line transform. We then merge the hough lines and create a "skeleton" of the pipe. We find phi_e and y_e and send it to the servers.